v2.2.8d

+++ RobloxFastFlagsInstaller.py
@@ -1,7 +1,7 @@
 # 
 # Roblox Fast Flags Installer
 # Made by Efaz from efaz.dev
-# v2.3.0
+# v2.3.1
 # 
 # Fulfill your Roblox needs and configuration through Python!
 # 
@@ -29,7 +29,7 @@ main_os = platform.system()
 cur_path = os.path.dirname(os.path.abspath(__file__))
 user_folder = (os.path.expanduser("~") if main_os == "Darwin" else os.getenv('LOCALAPPDATA'))
 orangeblox_mode = False
-script_version = "2.3.0"
+script_version = "2.3.1"
 def getLocalAppData():
     import platform
     import os
@@ -878,8 +878,8 @@ class pip:
                     downed_paths = []
                     for url_path_1 in url_paths:
                         url_path_2 = url_paths_2[co]
-                        self.requests.download(f"https://github.com/{url_path_2}/{url_path_1}/archive/refs/heads/main.zip", self._os.path.join(down_path, f"{url_path_1}.zip"))
-                        downed_paths.append(self._os.path.join(down_path, f"{url_path_1}.zip"))
+                        s = self.requests.download(f"https://github.com/{url_path_2}/{url_path_1}/archive/refs/heads/main.zip", self._os.path.join(down_path, f"{url_path_1}.zip"))
+                        if s.ok: downed_paths.append(self._os.path.join(down_path, f"{url_path_1}.zip"))
                         co += 1
                     return {"success": True, "path": down_path, "package_files": downed_paths}
                 else:
@@ -912,7 +912,7 @@ class pip:
                 with self._tempfile.NamedTemporaryFile(suffix=".py", delete=False) as temp_file: pypi_download_path = temp_file.name
                 if self.pythonSupported(3,9,0): download_res = self.requests.download("https://bootstrap.pypa.io/get-pip.py", pypi_download_path)      
                 else: current_python_version = self.getCurrentPythonVersion(); download_res = self.requests.download(f"https://bootstrap.pypa.io/pip/{current_python_version.split('.')[0]}.{current_python_version.split('.')[1]}/get-pip.py", pypi_download_path)
-                if download_res.returncode == 0:
+                if download_res.ok:
                     self.printDebugMessage(f"Successfully downloaded pip! Installing to Python..")
                     install_to_py = self._subprocess.run([self.executable, pypi_download_path], stdout=self.debug == False and self._subprocess.DEVNULL, stderr=self.debug == False and self._subprocess.DEVNULL)
                     if install_to_py.returncode == 0:
@@ -1039,7 +1039,7 @@ class pip:
             url = f"https://www.python.org/ftp/python/{version_url_folder}/python-{version}-macos11.pkg"
             with self._tempfile.NamedTemporaryFile(suffix=".pkg", delete=False) as temp_file: pkg_file_path = temp_file.name
             result = self.requests.download(url, pkg_file_path)            
-            if result.returncode == 0:
+            if result.ok:
                 self._subprocess.run(["open", pkg_file_path], stdout=self.debug == False and self._subprocess.DEVNULL, stderr=self.debug == False and self._subprocess.DEVNULL, check=True)
                 while self.getIfProcessIsOpened("/System/Library/CoreServices/Installer.app") == True: self._time.sleep(0.1)
                 self.printDebugMessage(f"Python installer has been executed: {pkg_file_path}")
@@ -1052,7 +1052,7 @@ class pip:
             else: url = f"https://www.python.org/ftp/python/{version_url_folder}/python-{version}.exe"
             with self._tempfile.NamedTemporaryFile(suffix=".exe", delete=False) as temp_file: exe_file_path = temp_file.name
             result = self.requests.download(url, exe_file_path)
-            if result.returncode == 0:
+            if result.ok:
                 self._subprocess.run([exe_file_path], stdout=self.debug == False and self._subprocess.DEVNULL, stderr=self.debug == False and self._subprocess.DEVNULL, check=True)
                 self.printDebugMessage(f"Python installer has been executed: {exe_file_path}")
             else:
@@ -3449,11 +3449,14 @@ class Handler:
                         if debug == True: printDebugMessage(f"Downloading Roblox {client_label} DMG from Roblox's servers..")
                         cur_vers_down_link = f'https://{self.getBestRobloxDownloadServer()}/{starter_url}mac/{"arm64/" if platform.machine() == "arm64" else ""}{cur_vers.get("client_version")}-Roblox{"Studio" if studio == True else ""}.zip'
                         if debug == True: printDebugMessage(f"Downloading from: {cur_vers_down_link}")
-                        requests.download(cur_vers_down_link, os.path.join(cur_path, f"Roblox{client_label}Install.zip"))
-                        zip_extract = pip_class.unzipFile(os.path.join(cur_path, f"Roblox{client_label}Install.zip"), filePath, ["Contents"])
-                        if zip_extract.returncode == 0: os.remove(os.path.join(cur_path, f"Roblox{client_label}Install.zip"))
+                        down_req = requests.download(cur_vers_down_link, os.path.join(cur_path, f"Roblox{client_label}Install.zip"))
+                        if down_req.ok:
+                            zip_extract = pip_class.unzipFile(os.path.join(cur_path, f"Roblox{client_label}Install.zip"), filePath, ["Contents"])
+                            if zip_extract.returncode == 0: os.remove(os.path.join(cur_path, f"Roblox{client_label}Install.zip"))
+                            else:
+                                if debug == True: printDebugMessage(f"Unable to unzip Roblox {client_label} installer due to an error.")
                         else:
-                            if debug == True: printDebugMessage(f"Unable to unzip Roblox {client_label} installer due to an error.")
+                            if debug == True: printDebugMessage(f"Unable to download Roblox {client_label} installer due to an error. Code: {down_req.status_code}")
                     else:
                         if debug == True: printDebugMessage(f"Unable to download Roblox {client_label} installer due to an http error.")
                 elif self.__main_os__ == "Windows":
@@ -3462,9 +3465,12 @@ class Handler:
                         if debug == True: printDebugMessage(f"Downloading Roblox EXE from Roblox's servers..")
                         cur_vers_down_link = f'https://{self.getBestRobloxDownloadServer()}/{starter_url}{cur_vers.get("client_version")}-Roblox{client_label}Installer.exe'
                         if debug == True: printDebugMessage(f"Downloading from: {cur_vers_down_link}")
-                        requests.download(cur_vers_down_link, filePath)
-                        if debug == True: printDebugMessage(f"Successfully downloaded installer!")
-                        return filePath
+                        down_req = requests.download(cur_vers_down_link, filePath)
+                        if down_req.ok:
+                            if debug == True: printDebugMessage(f"Successfully downloaded installer!")
+                            return filePath
+                        else:
+                            if debug == True: printDebugMessage(f"Unable to download Roblox {client_label} installer due to an http error. Code: {down_req.status_code}")
                     else:
                         if debug == True: printDebugMessage(f"Unable to download Roblox {client_label} installer due to an http error.")
                 else: self.unsupportedFunction()
@@ -4012,8 +4018,14 @@ class Handler:
                                             if not i == "":
                                                 if submit_status: submit_status.submit(f"[BUNDLE] Downloading Package [{i}]..", round((per_step/(len(marked_install_files)))*100, 2))
                                                 if debug == True: printDebugMessage(f"Downloading from Roblox's server: {i} [{round((per_step/(len(marked_install_files)))*100, 2)}/100]")
-                                                requests.download(f'https://{self.getBestRobloxDownloadServer()}/{starter_url}{cur_vers.get("client_version")}-{i}', os.path.join(installPath, i))
-                                                downloaded_zip_files.append(i)
+                                                down_req = requests.download(f'https://{self.getBestRobloxDownloadServer()}/{starter_url}{cur_vers.get("client_version")}-{i}', os.path.join(installPath, i))
+                                                if down_req.ok: downloaded_zip_files.append(i)
+                                                else:
+                                                    printErrorMessage(f"Unable to install Roblox due to a download error.")
+                                                    if alleged_path and os.path.exists(alleged_path): os.remove(alleged_path)
+                                                    if submit_status: submit_status.submit(f"\033ERR[BUNDLE] Unable to install Roblox due to a download error.", 80)
+                                                    if os.path.exists(installPath): shutil.rmtree(installPath, ignore_errors=True)
+                                                    return {"success": False}
                                         if verify == True:
                                             per_step = 0
                                             verified = True
@@ -4139,8 +4151,8 @@ class Handler:
                                 class download_stat:
                                     def submit(self, info):
                                         if submit_status: submit_status.submit(f"[BUNDLE] Downloading Roblox App!", int((info.percent/10)*3))
-                                requests.download(roblox_player_down, os.path.join(installPath, zip_name), submit_status=download_stat())
-                                if os.path.exists(os.path.join(installPath, zip_name)):
+                                down_req = requests.download(roblox_player_down, os.path.join(installPath, zip_name), submit_status=download_stat())
+                                if down_req.ok and os.path.exists(os.path.join(installPath, zip_name)):
                                     if os.path.exists(os.path.join(installPath, f"Roblox{client_label}")) or os.path.exists(appPath):
                                         if debug == True: printDebugMessage(f"Cleaning before install..")
                                         if os.path.exists(os.path.join(installPath, f"Roblox{client_label}")): shutil.rmtree(os.path.join(installPath, f"Roblox{client_label}"), ignore_errors=True)
@@ -4169,6 +4181,7 @@ class Handler:
                                     if debug == True: printDebugMessage(f"Unable to download the Roblox {client_label}.")
                                     if submit_status: submit_status.submit(f"\033ERR[BUNDLE] Failed to download Roblox {client_label}.", 100)
                                     if os.path.exists(appPath): shutil.rmtree(appPath, ignore_errors=True)
+                                    if os.path.exists(os.path.join(installPath, zip_name)): os.remove(os.path.join(installPath, zip_name))
                                     if alleged_path and os.path.exists(alleged_path): os.remove(alleged_path)
                                     return {"success": False}
                             except Exception as e:
